requirements.txt
langchain 
langchain-openai 
langchain-groq
pandas
ollama
python-dotenv
elasticsearch
rank-bm25
chromadb
fastembed
qdrant-client